requirements.txt
langchain==0.3.25
langchain_community==0.3.24
langchain_core==0.3.61
langchain_openai==0.3.18
numpy==2.2.6
pandas==2.2.3
psycopg2_binary==2.9.10
python-dotenv==1.1.0
